sysctl.d/cachyos-settings: Remove kernel.sched_rt_runtime_us (#111)
The comments here makes this tunable a bit misleading. Runtime is actually the allocated time from the total
period time that is given for RT tasks. A value of -1 indicates that the runtime is equal to the total period, i.e.
no limit from total period, and no extra time for non-RT tasks. Reading the documentation[1], it seems that
the kernel defaults are favorable compared to what we have right now.

[1] https://github.com/CachyOS/linux/blob/6.12/fixes/Documentation/scheduler/sched-rt-group.rst#21-system-wide-settings

Signed-off-by: Eric Naim <dnaim@cachyos.org>