sysctl: Remove net.ipv4.tcp_timestamps
This should not be set at a distribution level as it is a known security level. The benefits of disabling this is only
present on very high connection speeds. As users that have very high speeds are the minority, it strengthens the argument
that this should be set per-user rather than by the distribution.

Signed-off-by: Eric Naim <dnaim@cachyos.org>