sysctl: Reduce net.core.netdev_max_backlog value
In most cases, the default 1024 should already be sufficient. A higher value only benefits very fast connection speeds.
Even then, I believe that a value of 16384 is overkill. Set this to a saner value of 4096, still considered high but
it should be sufficient to prevent packet loss on those speedy connection speeds.

Signed-off-by: Eric Naim <dnaim@cachyos.org>