sbctl-batch-sign: Filter for unsigned files
Previously, the script was just grabbing all files recognized by sbctl and mass signing them. This was used
to check if the script was working correctly (i.e. signing files with sbctl) but wasn't changed to filter
for unsigned files instead after checking it worked correctly.

I left it untouched because it didn't seem like a problem at the time, but issues are arising now due to
this poorly written script.

Signed-off-by: Eric Naim <dnaim@cachyos.org>